requirements.txt
pip
streamlit
pandas
numpy
python-dotenv
ipykernel
pip
streamlit_chat
streamlit_pills
trl
python-dotenv